refactor: reorganize utils to reduce dependency contagion
Split monolithic util packages into focused subpackages so consumers
only pull the dependencies they actually need.

- pkg/utils/common.go: extract karpenter scheme registration to
  pkg/utils/karpenter/, and SKU/GPU helpers (GetSKUHandler,
  GetGPUConfigFromNodeLabels, etc.) to pkg/sku/. common.go no longer
  imports karpenter-provider-aws, sigs.k8s.io/karpenter, or pkg/apis.

- pkg/utils/resources/: split node helpers (GetNode, ListNodes,
  CheckNvidiaPlugin, NodeIsReadyAndNotDeleting, nvidia label/capacity
  constants) into pkg/utils/nodes/. pkg/utils/resources/ retains only
  the manifest CRUD helpers that pull in fluxcd types.

- pkg/workspace/inference/max_model_len_strategy.go: move to
  pkg/workspace/estimator/ so the estimator package stops depending
  on the inference package.

Net effect: pkg/workspace/estimator/nodesestimator no longer
transitively depends on fluxcd, helm-controller, source-controller,
or the inference package. Verified via 'go list -deps'.
