+++ api/v1alpha1/workspace_validation.go
@@ -359,7 +359,7 @@ func (r *ResourceSpec) validateCreateWithInference(inference *InferenceSpec, byp
 	}
 	instanceType := string(r.InstanceType)
 
-	skuHandler, err := utils.GetSKUHandler()
+	skuHandler, err := sku.GetSKUHandler()
 	if err != nil {
 		errs = errs.Also(apis.ErrGeneric(fmt.Sprintf("Failed to get SKU handler: %v", err), "instanceType"))
 		return errs
