+++ pkg/utils/nodes/nodes.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package resources
+package nodes
 
 import (
 	"context"
