+++ pkg/workspace/estimator/max_model_len.go
@@ -1,0 +1,154 @@
+// Copyright (c) KAITO authors.
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package estimator
+
+import (
+	"strings"
+
+	"k8s.io/apimachinery/pkg/api/resource"
+	"k8s.io/klog/v2"
+)
+
+// MaxModelLenInput contains the inputs required to compute the optimal
+// max model length for GPU memory efficiency. The struct is intentionally
+// self-contained so that external callers do not have to depend on any
+// internal Kaito model or SKU types.
+type MaxModelLenInput struct {
+	// ModelTokenLimit is the upper bound on tokens accepted by the model
+	// (typically max_position_embeddings from the model config).
+	ModelTokenLimit int
+	// BytesPerToken is the per-token KV cache footprint, in bytes.
+	BytesPerToken int
+	// TotalModelWeightSize is a Kubernetes resource.Quantity formatted string
+	// representing the total on-disk safetensor weight size (e.g. "25Gi").
+	TotalModelWeightSize string
+	// DisableTensorParallelism indicates the model does not shard weights
+	// or KV cache across GPUs/nodes (e.g. Falcon-style models).
+	DisableTensorParallelism bool
+	// GPUMemoryBytes is the memory of a single GPU, in bytes.
+	GPUMemoryBytes int64
+	// GPUCount is the number of GPUs per node.
+	GPUCount int
+	// NumRequiredNodes is the number of nodes serving the model.
+	NumRequiredNodes int
+}
+
+// ComputeMaxModelLen calculates the optimal max model length for GPU memory
+// efficiency given the provided model, GPU, and topology inputs.
+//
+// Returns 0 when the inputs are invalid or insufficient memory is available.
+func ComputeMaxModelLen(in MaxModelLenInput) int {
+	if in.ModelTokenLimit <= 0 || in.BytesPerToken <= 0 ||
+		in.GPUMemoryBytes <= 0 || in.GPUCount <= 0 || in.NumRequiredNodes <= 0 {
+		return 0
+	}
+
+	weightGiB, ok := parseModelWeight(in.TotalModelWeightSize)
+	if !ok {
+		return 0
+	}
+
+	availableBytes, adjustedBytesPerToken := calculateMemoryParameters(in, weightGiB)
+	if availableBytes <= 0 || adjustedBytesPerToken <= 0 {
+		return 0
+	}
+
+	candidate := int(availableBytes / adjustedBytesPerToken)
+	if candidate <= 0 {
+		return 0
+	}
+
+	finalResult := applyConstraintsAndAlignment(candidate, in.ModelTokenLimit)
+
+	klog.Infof("ComputeMaxModelLen: final result=%d", finalResult)
+	return finalResult
+}
+
+// parseModelWeight parses the model weight string and returns the weight in GiB.
+// It handles Kubernetes resource.Quantity format strings like "25.63Gi".
+func parseModelWeight(totalModelWeightSize string) (float64, bool) {
+	s := strings.TrimSpace(totalModelWeightSize)
+	if s == "" {
+		return 0, false
+	}
+
+	quantity, err := resource.ParseQuantity(s)
+	if err != nil {
+		return 0, false
+	}
+
+	bytes := quantity.Value()
+	if bytes <= 0 {
+		return 0, false
+	}
+
+	weightGiB := float64(bytes) / (1 << 30)
+	if weightGiB <= 0 {
+		return 0, false
+	}
+
+	return weightGiB, true
+}
+
+// calculateMemoryParameters computes available GPU memory and adjusted bytes per token.
+// Returns the available memory in bytes and the adjusted bytes per token for the calculation.
+func calculateMemoryParameters(in MaxModelLenInput, weightGiB float64) (float64, float64) {
+	gpuMemGB := float64(in.GPUMemoryBytes) / (1 << 30)
+	gpuCount := float64(in.GPUCount)
+	nodes := float64(in.NumRequiredNodes)
+	bytesPerToken := float64(in.BytesPerToken)
+
+	// availableMemoryGiB = (gpuMemGB * 0.84) / gpuCount - (weightGiB * 1.02) / (nodes * gpuCount) - 2.3
+
+	// Available GPU memory per GPU with 84% utilization factor.
+	usableMemoryPerGPU := (gpuMemGB * 0.84) / gpuCount
+
+	// Model weight overhead per GPU with 2% safety margin.
+	var modelWeightOverhead float64
+	if in.DisableTensorParallelism {
+		// Falcon-style models: don't distribute weight across nodes/GPUs.
+		modelWeightOverhead = weightGiB * 1.02
+	} else {
+		modelWeightOverhead = (weightGiB * 1.02) / (nodes * gpuCount)
+	}
+
+	// Static overhead for activations and non-torch components
+	// (max activations 1.7 GiB + max non-torch overhead 0.6 GiB).
+	staticOverhead := 2.3
+
+	availableMemoryGiB := usableMemoryPerGPU - modelWeightOverhead - staticOverhead
+	availableMemoryBytes := availableMemoryGiB * (1 << 30)
+
+	var adjustedBytesPerToken float64
+	if in.DisableTensorParallelism {
+		adjustedBytesPerToken = bytesPerToken
+	} else {
+		adjustedBytesPerToken = bytesPerToken / gpuCount * nodes
+	}
+
+	return availableMemoryBytes, adjustedBytesPerToken
+}
+
+// applyConstraintsAndAlignment clamps candidate to modelTokenLimit and aligns
+// down to a 256-token boundary for efficient memory allocation.
+func applyConstraintsAndAlignment(candidate, modelTokenLimit int) int {
+	originalCandidate := candidate
+	if modelTokenLimit > 0 && candidate > modelTokenLimit {
+		candidate = modelTokenLimit
+		klog.Infof("ComputeMaxModelLen: clamped to ModelTokenLimit: %d -> %d", originalCandidate, candidate)
+	}
+
+	candidate = (candidate / 256) * 256
+	return candidate
+}
