+++ pkg/workspace/estimator/nodesestimator/estimator.go
@@ -24,9 +24,8 @@ import (
 
 	kaitov1beta1 "github.com/kaito-project/kaito/api/v1beta1"
 	"github.com/kaito-project/kaito/pkg/sku"
-	"github.com/kaito-project/kaito/pkg/utils"
 	"github.com/kaito-project/kaito/pkg/utils/consts"
-	"github.com/kaito-project/kaito/pkg/utils/resources"
+	"github.com/kaito-project/kaito/pkg/utils/nodes"
 	estimator "github.com/kaito-project/kaito/pkg/workspace/estimator"
 	"github.com/kaito-project/kaito/presets/workspace/models"
 )
@@ -59,26 +58,26 @@ func (c *NodeEstimator) EstimateNodeCount(ctx context.Context, req estimator.Nod
 	if req.ResourceProfile.DisableNodeAutoProvisioning {
 		// NAP is disabled (BYO scenario) — derive GPU config from existing ready nodes.
 		matchLabels := client.MatchingLabels(kaitov1beta1.SanitizedMatchLabels(req.ResourceProfile.LabelSelector))
-		nodeList, listErr := resources.ListNodes(ctx, cl, matchLabels)
+		nodeList, listErr := nodes.ListNodes(ctx, cl, matchLabels)
 		if listErr != nil {
 			return 0, fmt.Errorf("failed to list ready nodes: %w", listErr)
 		}
 		var readyNodes []*corev1.Node
 		for i := range nodeList.Items {
-			if resources.NodeIsReadyAndNotDeleting(&nodeList.Items[i]) {
+			if nodes.NodeIsReadyAndNotDeleting(&nodeList.Items[i]) {
 				readyNodes = append(readyNodes, &nodeList.Items[i])
 			}
 		}
 		if len(readyNodes) == 0 {
 			return 0, fmt.Errorf("no ready nodes found, unable to determine GPU configuration")
 		}
-		gpuConfig, err = utils.GetGPUConfigFromNodeLabels(readyNodes[0])
+		gpuConfig, err = sku.GetGPUConfigFromNodeLabels(readyNodes[0])
 		if err != nil {
 			return 0, fmt.Errorf("failed to get GPU config from existing nodes: %w", err)
 		}
 	} else {
 		// NAP is enabled — instanceType is required and must be valid.
-		gpuConfig, err = utils.GetGPUConfigBySKU(req.ResourceProfile.InstanceType)
+		gpuConfig, err = sku.GetGPUConfigBySKU(req.ResourceProfile.InstanceType)
 		if err != nil {
 			return 0, fmt.Errorf("failed to get GPU config for instance type %s: %w", req.ResourceProfile.InstanceType, err)
 		}
