+++ pkg/workspace/inference/preset_inferences.go
@@ -39,6 +39,8 @@ import (
 	"github.com/kaito-project/kaito/pkg/utils/consts"
 	"github.com/kaito-project/kaito/pkg/utils/generator"
 	"github.com/kaito-project/kaito/pkg/utils/resources"
+	"github.com/kaito-project/kaito/pkg/utils/nodes"
+	"github.com/kaito-project/kaito/pkg/workspace/estimator"
 	"github.com/kaito-project/kaito/pkg/workspace/manifests"
 	metadata "github.com/kaito-project/kaito/presets/workspace/models"
 )
@@ -87,7 +89,7 @@ func defaultTolerations(ws *v1beta1.Workspace) []corev1.Toleration {
 		{
 			Effect:   corev1.TaintEffectNoSchedule,
 			Operator: corev1.TolerationOpExists,
-			Key:      resources.CapacityNvidiaGPU,
+			Key:      nodes.CapacityNvidiaGPU,
 		},
 		{
 			Effect:   corev1.TaintEffectNoSchedule,
@@ -97,7 +99,7 @@ func defaultTolerations(ws *v1beta1.Workspace) []corev1.Toleration {
 		},
 	}
 
-	if utils.IsAzureCloudProvider() {
+	if sku.IsAzureCloudProvider() {
 		tolerations = append(tolerations, corev1.Toleration{
 			Effect:   corev1.TaintEffectNoSchedule,
 			Key:      consts.SpotInstanceKey,
@@ -210,7 +212,7 @@ func getGPUConfig(ctx *generator.WorkspaceGeneratorContext) (*sku.GPUConfig, err
 	if featuregates.FeatureGates[consts.FeatureFlagDisableNodeAutoProvisioning] {
 		// NAP is disabled (BYO scenario) - prefer to get GPU config from matching nodes with nvidia.com labels
 		// Only try to find matching nodes if we have a labelSelector and if WorkerNodes is not already populated
-		readyNodes, err := resources.GetReadyNodes(ctx.Ctx, ctx.KubeClient, ctx.Workspace)
+		readyNodes, err := nodes.GetReadyNodes(ctx.Ctx, ctx.KubeClient, ctx.Workspace)
 		if err != nil {
 			return nil, fmt.Errorf("failed to list ready nodes: %w", err)
 		}
@@ -218,10 +220,10 @@ func getGPUConfig(ctx *generator.WorkspaceGeneratorContext) (*sku.GPUConfig, err
 			return nil, fmt.Errorf("no ready nodes found matching the workspace's label selector")
 		}
 
-		return utils.GetGPUConfigFromNodeLabels(readyNodes[0])
+		return sku.GetGPUConfigFromNodeLabels(readyNodes[0])
 	} else {
 		// NAP is enabled - try to get GPU config from known SKU
-		gpuConfig, err := utils.GetGPUConfigBySKU(ctx.Workspace.Resource.InstanceType)
+		gpuConfig, err := sku.GetGPUConfigBySKU(ctx.Workspace.Resource.InstanceType)
 		if err != nil {
 			return nil, err
 		}
@@ -422,10 +424,10 @@ func GenerateInferencePodSpec(gpuConfig *sku.GPUConfig, numNodes int) func(*gene
 		// resource requirements
 		resourceReq := corev1.ResourceRequirements{
 			Requests: corev1.ResourceList{
-				corev1.ResourceName(resources.CapacityNvidiaGPU): *resource.NewQuantity(int64(gpuConfig.GPUCount), resource.DecimalSI),
+				corev1.ResourceName(nodes.CapacityNvidiaGPU): *resource.NewQuantity(int64(gpuConfig.GPUCount), resource.DecimalSI),
 			},
 			Limits: corev1.ResourceList{
-				corev1.ResourceName(resources.CapacityNvidiaGPU): *resource.NewQuantity(int64(gpuConfig.GPUCount), resource.DecimalSI),
+				corev1.ResourceName(nodes.CapacityNvidiaGPU): *resource.NewQuantity(int64(gpuConfig.GPUCount), resource.DecimalSI),
 			},
 		}
 
@@ -446,7 +448,15 @@ func GenerateInferencePodSpec(gpuConfig *sku.GPUConfig, numNodes int) func(*gene
 						//klog.Infof("[RuntimeContext] workspace=%s using user explicit max-model-len=%d", ctx.Workspace.Name, maxModelLen)
 						//} else {
 						// If no user value, compute planned value
-						maxModelLen = computeMaxModelLen(presetParams, gpuConfig, numNodes)
+						maxModelLen = estimator.ComputeMaxModelLen(estimator.MaxModelLenInput{
+							ModelTokenLimit:          presetParams.ModelTokenLimit,
+							BytesPerToken:            presetParams.BytesPerToken,
+							TotalModelWeightSize:     presetParams.TotalSafeTensorFileSize,
+							DisableTensorParallelism: presetParams.DisableTensorParallelism,
+							GPUMemoryBytes:           gpuConfig.GPUMem.Value(),
+							GPUCount:                 gpuConfig.GPUCount,
+							NumRequiredNodes:         numNodes,
+						})
 						klog.Infof("[RuntimeContext] workspace=%s using computed max-model-len=%d (gpuConfig=%s, numNodes=%d)", ctx.Workspace.Name, maxModelLen, gpuConfig.String(), numNodes)
 						//}
 					}
