+++ test/e2e/utils/cluster.go
@@ -27,7 +27,7 @@ import (
 
 	kaitov1alpha1 "github.com/kaito-project/kaito/api/v1alpha1"
 	kaitov1beta1 "github.com/kaito-project/kaito/api/v1beta1"
-	kaitoutils "github.com/kaito-project/kaito/pkg/utils"
+	karpenterutils "github.com/kaito-project/kaito/pkg/utils/karpenter"
 )
 
 const (
@@ -57,9 +57,9 @@ func GetClusterClient(cluster *Cluster) {
 	utilruntime.Must(clientgoscheme.AddToScheme(scheme))
 	utilruntime.Must(kaitov1alpha1.AddToScheme(scheme))
 	utilruntime.Must(kaitov1beta1.AddToScheme(scheme))
-	utilruntime.Must(kaitoutils.KarpenterSchemeBuilder.AddToScheme(scheme))
+	utilruntime.Must(karpenterutils.KarpenterSchemeBuilder.AddToScheme(scheme))
 	utilruntime.Must(azurev1beta1.SchemeBuilder.AddToScheme(scheme))
-	utilruntime.Must(kaitoutils.AwsSchemeBuilder.AddToScheme(scheme))
+	utilruntime.Must(karpenterutils.AwsSchemeBuilder.AddToScheme(scheme))
 	utilruntime.Must(helmv2.AddToScheme(scheme))
 	utilruntime.Must(sourcev1.AddToScheme(scheme))
